% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matches.R
\name{bind_matches}
\alias{bind_matches}
\title{Bind Match Indicator Columns to a Data Frame}
\usage{
bind_matches(.df, ...)
}
\arguments{
\item{.df}{A data frame.}

\item{...}{\code{matchit} objects returned by the \code{MatchIt} package. They can be
named or unnamed.}
}
\value{
\code{.df} with addition columns for every element of \code{...}
}
\description{
Given a data frame \code{.df}, the function \code{bind_matches} creates binary
indicator variables for each match returned by the \code{MatchIt} library and
binds the resulting columns to \code{.df}. In other words, the result is the
original data frame plus a column for however many matches you want to bind.
}
