% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_love.R
\name{geom_love}
\alias{geom_love}
\title{Create a Love plot}
\usage{
geom_love(
  line_size = 0.8,
  point_size = 1.85,
  vline_xintercept = 0.1,
  vline_color = "grey70",
  vline_size = 0.6
)
}
\arguments{
\item{line_size}{The line size, passed to \code{\link[ggplot2:geom_path]{ggplot2::geom_line()}}.}

\item{point_size}{The point size, passed to \code{\link[ggplot2:geom_point]{ggplot2::geom_point()}}.}

\item{vline_xintercept}{The X intercept, passed to \code{\link[ggplot2:geom_abline]{ggplot2::geom_vline()}}.}

\item{vline_color}{The vertical line color, passed to
\code{\link[ggplot2:geom_abline]{ggplot2::geom_vline()}}.}

\item{vline_size}{The vertical line size, passed to
\code{\link[ggplot2:geom_abline]{ggplot2::geom_vline()}}.}
}
\value{
a list of \code{geoms}.
}
\description{
\code{geom_love()} is a helper function to create Love plots in ggplot2. Love
plots are a diagnostic approach to assessing balance before and after
weighting. Many researchers use 0.1 on the absolute SMD scale to evaluate if
a variable is well-balanced between groups, although this is just a rule of
thumb. \code{geom_love()} is a simple wrapper around \code{\link[ggplot2:geom_point]{ggplot2::geom_point()}},
\code{\link[ggplot2:geom_path]{ggplot2::geom_line()}}, and \code{\link[ggplot2:geom_abline]{ggplot2::geom_vline()}}. It also adds default
aesthetics via \code{\link[ggplot2:aes]{ggplot2::aes()}}. For more complex Love plots, we recommend
using ggplot2 directly.
}
\examples{
library(ggplot2)
plot_df <- tidy_smd(
  nhefs_weights,
  race:active,
  .group = qsmk,
  .wts = starts_with("w_")
)

# no need for `aes()` by default.
ggplot(plot_df) + geom_love()

}
