% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidysmd-package.R
\docType{package}
\name{tidysmd-package}
\alias{tidysmd}
\alias{tidysmd-package}
\title{tidysmd: Tidy Standardized Mean Differences}
\description{
Tidy standardized mean differences ('SMDs'). 'tidysmd' uses the 'smd' package to calculate standardized mean differences for variables in a dataframe, returning the results in a tidy format.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/malcolmbarrett/tidysmd}
  \item Report bugs at \url{https://github.com/malcolmbarrett/tidysmd/issues}
}

}
\author{
\strong{Maintainer}: Malcolm Barrett \email{malcolmbarrett@gmail.com} (\href{https://orcid.org/0000-0003-0299-5825}{ORCID})

}
\keyword{internal}
